rename module test to testmock
