+++ server/testmock/pokeshop.go
@@ -1,4 +1,4 @@
-package test
+package testmock
 
 import (
 	"context"
